[WP] Add Go pprof-label span tester as its own binary
