+++ pkg/security/tests/syscall_tester/.gitignore
@@ -1,5 +1,6 @@
 go/ebpf_probe.o
 bin/syscall_go_tester
+bin/span_go_tester
 bin/syscall_tester
 bin/syscall_x86_tester
 bin/syscall_tester.d
